refactor(cycle): peel synthesize depth/budget helpers into synthesize-budget.ts; satisfy verify gates
- Move the pure depth-tier and chunk-budget helpers (prepareTranscriptForSynthesis,
  inferSynthesisDepth, triageDepthHint, computeChunkCharBudget, depth budget
  resolution, prompt-block and telemetry builders) into
  src/core/cycle/synthesize-budget.ts. synthesize.ts re-exports the symbols
  tests and dream-retriage import, so callers are unchanged. Brings
  synthesize.ts from 3398 to 3194 lines, back under its 3216 ratchet ceiling.
- Raise the autopilot.ts (+2) and sync.ts (+10) ceilings for the small
  since/until and dream-budget wiring, with reviewer notes in the tsv.
- Regenerate scripts/structural-suites.tsv (bun scripts/classify-tests.ts).
- KEY_FILES.md: entry for the new module; synthesize.ts entry points at it.

bun run verify: 54/54 green. bun run typecheck clean.
